rename(components): display 폴더 생성
