+++ apps/monitor-web/src/components/common/buttons/index.ts
@@ -1,3 +1,4 @@
 export { default as BasicButton } from "./BasicButton";
+export { default as Checkbox } from "./CheckboxButton";
 export { default as ClearButton } from "./ClearButton";
 export { default as IconButton } from "./IconButton";
