+++ apps/monitor-web/src/components/common/display/Icon.tsx
