+++ apps/monitor-web/src/components/common/display/Image.tsx
