+++ apps/monitor-web/src/components/common/display/index.ts
@@ -1,0 +1,3 @@
+export { default as Icon } from "./Icon";
+export type { IconName } from "./Icon";
+export { default as Image } from "./Image";
