+++ apps/monitor-web/src/components/common/index.ts
@@ -1,9 +1,5 @@
 export * from "./buttons";
+export * from "./display";
 export * from "./feedback";
 export * from "./inputs";
 export * from "./modals";
-
-export { default as Checkbox } from "./Checkbox";
-export { default as Icon } from "./Icon";
-export type { IconName } from "./Icon";
-export { default as Image } from "./Image";
